Updated previous msvc warning fixes

+++ Common/GameData/Character.cpp
@@ -247,7 +247,7 @@ void Character::GetCharBuildInfo(BitStream &src)
 
 void Character::sendEnhancements(BitStream &bs) const
 {
-    bs.StorePackedBits(5, m_char_data.m_enhancements.size()); // count
+    bs.StorePackedBits(5, (uint32_t)m_char_data.m_enhancements.size()); // count
     for(size_t i = 0; i < m_char_data.m_enhancements.size(); ++i)
     {
         bs.StorePackedBits(3, m_char_data.m_enhancements[i].m_slot_idx); // boost idx, maybe use m_enhancement_idx
@@ -283,11 +283,11 @@ void Character::sendInspirations(BitStream &bs) const
 
 void Character::sendOwnedPowers(BitStream &bs) const
 {
-    bs.StorePackedBits(4, m_char_data.m_powersets.size()); // count
+    bs.StorePackedBits(4, (uint32_t)m_char_data.m_powersets.size()); // count
     for(const CharacterPowerSet &pset : m_char_data.m_powersets)
     {
         bs.StorePackedBits(5, pset.m_level_bought);
-        bs.StorePackedBits(4, pset.m_powers.size());
+        bs.StorePackedBits(4, (uint32_t)pset.m_powers.size());
         for(const CharacterPower &power : pset.m_powers)
         {
             power.m_power_info.serializeto(bs);
@@ -297,7 +297,7 @@ void Character::sendOwnedPowers(BitStream &bs) const
             if(power.m_total_eh_slots > power.m_enhancements.size())
                 sCWarning(logPowers) << "sendOwnedPowers: Total EH Slots larger than vector!";
 
-            bs.StorePackedBits(4, power.m_enhancements.size());
+            bs.StorePackedBits(4, (uint32_t)power.m_enhancements.size());
             for(const CharacterEnhancement &eh : power.m_enhancements)
             {
                 bs.StoreBits(1, eh.m_slot_used); // slot has enhancement
@@ -485,10 +485,10 @@ void serializeStats(const Parse_CharAttrib &src,BitStream &bs, bool /*sendAbsolu
 {
     bs.StoreBits(1,1); // we have more data
     bs.StorePackedBits(1,0);
-    bs.StorePackedBits(5,src.m_HitPoints/5.0f);
+    bs.StorePackedBits(5, (uint32_t)(src.m_HitPoints / 5.0f));
     bs.StoreBits(1,1); // we have more data
     bs.StorePackedBits(1,1);
-    bs.StorePackedBits(5,src.m_Endurance/5.0f);
+    bs.StorePackedBits(5, (uint32_t)(src.m_Endurance / 5.0f));
     bs.StoreBits(1,0); // no more data
 }
 
@@ -496,10 +496,10 @@ void serializeFullStats(const Parse_CharAttrib &src,BitStream &bs, bool /*sendAb
 {
     bs.StoreBits(1,1); // we have more data
     bs.StorePackedBits(1,0);
-    bs.StorePackedBits(7,src.m_HitPoints);
+    bs.StorePackedBits(7, (uint32_t)src.m_HitPoints);
     bs.StoreBits(1,1); // we have more data
     bs.StorePackedBits(1,1);
-    bs.StorePackedBits(7,src.m_Endurance);
+    bs.StorePackedBits(7, (uint32_t)src.m_Endurance);
     bs.StoreBits(1,0); // no more data
 }
 
@@ -610,7 +610,7 @@ void Character::sendFriendList(BitStream &bs) const
     bs.StorePackedBits(1,1); // v2 = force_update
     bs.StorePackedBits(1,fl->m_friends_count);
 
-    for(int i=0; i<fl->m_friends_count; ++i)
+    for(int i=0; i<(int)fl->m_friends_count; ++i)
     {
         bs.StoreBits(1,fl->m_has_friends); // if false, client will skip this iteration
         bs.StorePackedBits(1,fl->m_friends[i].m_db_id);
